refactor: Move instruction filters to their own file
